# Implementation pack/05-openapi.yaml
openapi: 3.0.3
info:
  title: AI CV Tailoring and Cover Letter Platform — Backend API
  description: >
    Draft specification for developer implementation. Covers CV ingestion,
    job post ingestion, matching, tailored CV generation, guided cover
    letter generation, and export. All generated content must carry
    evidence references — see evidenceReferences fields throughout.
  version: 0.1.0-draft
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /cover-letters:
    get:
      tags: [cover-letters]
      summary: List cover-letter workflows for the current user, most recent first
      description: >-
        Account-only — cover letters have always been account-only
        throughout this router, matching GET /cvs and GET /job-posts'
        precedent for their own list endpoints.
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of cover-letter workflows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverLetterWorkflowListResponse' }
  /cover-letters/start:
    post:
      tags: [cover-letters]
      summary: Start a guided cover letter workflow from CV and job post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cvId, jobPostId]
              properties:
                cvId: { type: string, format: uuid }
                jobPostId: { type: string, format: uuid }
                matchId: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverLetterWorkflow' }
  /cover-letters/{workflowId}/questions:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    get:
      tags: [cover-letters]
      summary: Retrieve the next question set
      responses:
        '200':
          description: Question set
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CoverLetterQuestion' }
        '404': { $ref: '#/components/responses/NotFound' }
  /cover-letters/{workflowId}/answers:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    post:
      tags: [cover-letters]
      summary: Submit user answers for the active step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  items:
                    type: object
                    required: [questionId, answerText]
                    properties:
                      questionId: { type: string, format: uuid }
                      answerText: { type: string }
      responses:
        '202':
          description: >
            Answers accepted. On the final step, the returned `status` is
            `generating`, not `draft_ready` — draft generation now runs
            asynchronously (Sprint 4, real LLM generation), not
            synchronously inside this request. Poll
            `GET /cover-letters/{workflowId}/draft` (returns 404 until the
            draft is actually ready) rather than assuming readiness from
            this response.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverLetterWorkflow' }
        '422':
          description: A required question at the current step was left unanswered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /cover-letters/{workflowId}/draft:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    get:
      tags: [cover-letters]
      summary: Retrieve the current cover letter draft
      responses:
        '200':
          description: Draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverLetterDraft' }
        '404': { $ref: '#/components/responses/NotFound' }
  /cover-letters/{workflowId}/regenerate:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    post:
      tags: [cover-letters]
      summary: Regenerate letter after user edits or new answers
      responses:
        '202':
          description: Regeneration queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcessingJobRef' }
  /cover-letters/{workflowId}/approve:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    post:
      tags: [cover-letters]
      summary: Mark letter as approved for export
      responses:
        '200':
          description: Letter approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverLetterDraft' }
        '409':
          description: Draft not in a state that can be approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------------
  # JOBS, AUDIT, EXPORTS
  # ---------------------------------------------------------------------
...
  /exports/cover-letter/{workflowId}:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    post:
      tags: [exports]
      summary: Request a docx export package for an approved cover letter
      responses:
        '202':
          description: Export queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportRequest' }
        '409':
          description: Draft is not approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    workflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string, format: uuid }
    limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, default: 0, minimum: 0 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: >
        Standard error envelope returned by every non-2xx response.
        `details` is free-form and endpoint-specific — populate it with
        whatever helps the caller correct the request (offending field,
        allowed values, etc.), omit it when there's nothing further to add.
      required: [status, code, message, timestamp, path]
      properties:
        status: { type: integer, description: "HTTP status code, duplicated in the body for logging convenience" }
        code: { type: string, description: "Stable machine-readable error code, e.g. VALIDATION_ERROR" }
        message: { type: string, description: "Human-readable explanation" }
        timestamp: { type: string, format: date-time }
        path: { type: string, description: "Request path that produced the error" }
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        status: 400
        code: VALIDATION_ERROR
        message: "Invalid file type. Only PDF and DOCX are supported."
        timestamp: "2026-08-03T10:00:00Z"
        path: "/api/v1/cvs"
        details:
          field: file
          reason: unsupported_mime_type
          allowed: ["application/pdf", "application/vnd.openxmlformats-officedocument.wordprocessingml.document"]
...
    CoverLetterWorkflow:
      type: object
      properties:
        id: { type: string, format: uuid }
        cvId: { type: string, format: uuid }
        jobPostId: { type: string, format: uuid }
        matchId: { type: string, format: uuid, nullable: true }
        currentStep: { type: integer }
        status: { type: string, enum: [awaiting_answers, generating, draft_ready, approved, archived, generation_failed] }
        questionSetVersion: { type: integer }
        createdAt: { type: string, format: date-time }
    CoverLetterWorkflowListItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        jobPostId: { type: string, format: uuid }
        jobTitle: { type: string, nullable: true }
        employer: { type: string, nullable: true }
        status: { type: string, enum: [awaiting_answers, generating, draft_ready, approved, archived, generation_failed] }
        currentStep: { type: integer }
        createdAt: { type: string, format: date-time }
    CoverLetterWorkflowListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CoverLetterWorkflowListItem' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    CoverLetterQuestion:
      type: object
      properties:
        id: { type: string, format: uuid }
        stepNumber: { type: integer }
        questionText: { type: string }
        questionCategory:
          type: string
          enum: [employer_interest, motivation, relevant_example, tone_preference, availability, clarification]
    CoverLetterDraft:
      type: object
      properties:
        id: { type: string, format: uuid }
        workflowId: { type: string, format: uuid }
        versionNumber: { type: integer }
        status: { type: string, enum: [pending, generated, user_edited, approved, archived, failed] }
        bodyText: { type: string }
        evidenceReferences:
          type: array
          items: { type: string }
          description: "Must be non-empty — same evidence-gate rule as tailored CV sections. May reference cv_experience_items.id, cv_project_items.id, cv_skill_items.id, or cover_letter_answers.id."
        promptVersion: { type: string, nullable: true, description: "Versioned prompt template used for this draft version." }
        modelId: { type: string, nullable: true, description: "Model/provider that generated this draft version — 'rules-based' when the deterministic fallback was used instead of a real LLM call." }
        createdAt: { type: string, format: date-time }
        approvedAt: { type: string, format: date-time, nullable: true }
      example:
        id: "cl_draft_1"
        workflowId: "wf_1234"
        versionNumber: 1
        status: generated
        bodyText: "Dear Hiring Manager,\n\nI am writing to express my interest in the Senior Software Engineer position at TechCorp..."
        evidenceReferences: ["exp_1", "answer_q2"]
        promptVersion: "v1"
        modelId: "gpt-4o-mini-2024-07-18"
        createdAt: "2026-08-03T10:50:00Z"
        approvedAt: null
...
    ProcessingJobRef:
      type: object
      properties:
        jobId: { type: string, format: uuid }
        status: { type: string, enum: [queued] }
...
    ExportRequest:
      type: object
      description: >
        Sprint 5. format resolves the format decision this schema originally
        left as a placeholder (see 01-implementation-plan.md open decisions):
        docx (primary), pdf (secondary, only after downloadedAt is set), zip
        (application packs — two independent docx files, not a merged doc).
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, processing, completed, failed] }
        sourceDraftType: { type: string, enum: [tailored_cv, cover_letter, application_pack] }
        fileReference: { type: string, nullable: true, description: "Relative download URL (GET /exports/{id}/download) once status is completed; null otherwise." }
        format: { type: string, enum: [docx, pdf, zip] }
        templateId: { type: string, nullable: true, description: "CV template layout used; null for cover-letter/pdf/zip exports." }
        downloadedAt: { type: string, format: date-time, nullable: true, description: "Set the first time the docx has been downloaded — gates PDF conversion." }
        derivedFromExportId: { type: string, format: uuid, nullable: true, description: "Set on format=pdf rows, pointing at the source docx export." }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...